Resolved issues from merging
